LPS-170700 Check permission properly, check regular roles permission first but if the object definition is restricted by account we should make sure that the user has permissions in their account roles

+++ modules/apps/object/object-service/src/main/java/com/liferay/object/service/impl/ObjectEntryServiceImpl.java
@@ -14,13 +14,16 @@
 
 package com.liferay.object.service.impl;
 
+import com.liferay.account.constants.AccountConstants;
+import com.liferay.account.model.AccountEntry;
 import com.liferay.account.service.AccountEntryLocalService;
 import com.liferay.object.configuration.ObjectConfiguration;
 import com.liferay.object.constants.ObjectActionKeys;
-import com.liferay.object.entry.permission.util.ObjectEntryPermissionUtil;
+import com.liferay.object.exception.ObjectDefinitionAccountEntryRestrictedException;
 import com.liferay.object.exception.ObjectEntryCountException;
 import com.liferay.object.model.ObjectDefinition;
 import com.liferay.object.model.ObjectEntry;
+import com.liferay.object.model.ObjectField;
 import com.liferay.object.service.ObjectFieldLocalService;
 import com.liferay.object.service.base.ObjectEntryServiceBaseImpl;
 import com.liferay.object.service.persistence.ObjectDefinitionPersistence;
@@ -29,12 +32,19 @@ import com.liferay.osgi.service.tracker.collections.map.ServiceTrackerMapFactory
 import com.liferay.petra.string.StringBundler;
 import com.liferay.portal.aop.AopService;
 import com.liferay.portal.configuration.metatype.bnd.util.ConfigurableUtil;
+import com.liferay.portal.kernel.dao.orm.QueryUtil;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.model.User;
+import com.liferay.portal.kernel.security.auth.PrincipalException;
 import com.liferay.portal.kernel.security.permission.ActionKeys;
 import com.liferay.portal.kernel.security.permission.resource.ModelResourcePermission;
 import com.liferay.portal.kernel.security.permission.resource.PortletResourcePermission;
 import com.liferay.portal.kernel.service.ServiceContext;
+import com.liferay.portal.kernel.util.ArrayUtil;
+import com.liferay.portal.kernel.util.ListUtil;
+import com.liferay.portal.kernel.util.MapUtil;
+import com.liferay.portal.kernel.util.StringUtil;
+import com.liferay.portal.kernel.workflow.WorkflowConstants;
 
 import java.io.Serializable;
 
@@ -103,12 +113,79 @@ public class ObjectEntryServiceImpl extends ObjectEntryServiceBaseImpl {
 	}
 
 	@Override
+	public void checkPermission(
+			String actionId, ObjectEntry objectEntry, long userId)
+		throws PortalException {
+
+		ObjectDefinition objectDefinition =
+			_objectDefinitionPersistence.findByPrimaryKey(
+				objectEntry.getObjectDefinitionId());
+
+		ModelResourcePermission<ObjectEntry> modelResourcePermission =
+			_modelResourcePermissionsServiceTrackerMap.getService(
+				objectDefinition.getClassName());
+
+		if (modelResourcePermission.contains(
+				getPermissionChecker(), objectEntry, actionId)) {
+
+			return;
+		}
+
+		if (!objectDefinition.isAccountEntryRestricted()) {
+			throw new PrincipalException.MustHavePermission(
+				getPermissionChecker(), modelResourcePermission.getModelName(),
+				objectEntry.getObjectEntryId(), actionId);
+		}
+
+		ObjectField objectField = _objectFieldLocalService.getObjectField(
+			objectDefinition.getAccountEntryRestrictedObjectFieldId());
+
+		long accountEntryId = MapUtil.getLong(
+			objectEntry.getValues(), objectField.getName());
+
+		if (accountEntryId == 0) {
+			return;
+		}
+
+		long[] accountEntryIds = ListUtil.toLongArray(
+			_accountEntryLocalService.getUserAccountEntries(
+				userId, AccountConstants.PARENT_ACCOUNT_ENTRY_ID_DEFAULT, null,
+				new String[] {
+					AccountConstants.ACCOUNT_ENTRY_TYPE_BUSINESS,
+					AccountConstants.ACCOUNT_ENTRY_TYPE_PERSON
+				},
+				WorkflowConstants.STATUS_APPROVED, QueryUtil.ALL_POS,
+				QueryUtil.ALL_POS),
+			AccountEntry::getAccountEntryId);
+
+		if (ArrayUtil.contains(accountEntryIds, accountEntryId)) {
+			return;
+		}
+
+		if (StringUtil.equals(actionId, ActionKeys.DELETE) ||
+			StringUtil.equals(actionId, ActionKeys.VIEW)) {
+
+			throw new ObjectDefinitionAccountEntryRestrictedException(
+				StringBundler.concat(
+					"User ", userId, " must have ", actionId,
+					" permission for ", objectDefinition.getClassName(),
+					objectEntry.getObjectEntryId()));
+		}
+
+		throw new ObjectDefinitionAccountEntryRestrictedException(
+			StringBundler.concat(
+				"The account entry ", accountEntryId,
+				" does not exist or the user ", userId,
+				" does not belong to it"));
+	}
+
+	@Override
 	public ObjectEntry deleteObjectEntry(long objectEntryId)
 		throws PortalException {
 
-		_checkPermissions(
-			objectEntryLocalService.getObjectEntry(objectEntryId),
-			ActionKeys.DELETE);
+		checkPermission(
+			ActionKeys.DELETE,
+			objectEntryLocalService.getObjectEntry(objectEntryId), getUserId());
 
 		return objectEntryLocalService.deleteObjectEntry(objectEntryId);
 	}
@@ -121,7 +198,7 @@ public class ObjectEntryServiceImpl extends ObjectEntryServiceBaseImpl {
 		ObjectEntry objectEntry = objectEntryLocalService.getObjectEntry(
 			externalReferenceCode, companyId, groupId);
 
-		_checkPermissions(objectEntry, ActionKeys.DELETE);
+		checkPermission(ActionKeys.DELETE, objectEntry, getUserId());
 
 		return objectEntryLocalService.deleteObjectEntry(objectEntry);
 	}
@@ -134,7 +211,8 @@ public class ObjectEntryServiceImpl extends ObjectEntryServiceBaseImpl {
 			objectEntryId);
 
 		if (objectEntry != null) {
-			_checkPermissions(objectEntry, ActionKeys.VIEW);
+			objectEntryService.checkPermission(
+				ActionKeys.VIEW, objectEntry, getUserId());
 		}
 
 		return objectEntry;
@@ -177,7 +255,8 @@ public class ObjectEntryServiceImpl extends ObjectEntryServiceBaseImpl {
 		ObjectEntry objectEntry = objectEntryLocalService.getObjectEntry(
 			objectEntryId);
 
-		_checkPermissions(objectEntry, ActionKeys.VIEW);
+		objectEntryService.checkPermission(
+			ActionKeys.VIEW, objectEntry, getUserId());
 
 		return objectEntry;
 	}
@@ -190,7 +269,8 @@ public class ObjectEntryServiceImpl extends ObjectEntryServiceBaseImpl {
 		ObjectEntry objectEntry = objectEntryLocalService.getObjectEntry(
 			externalReferenceCode, companyId, groupId);
 
-		_checkPermissions(objectEntry, ActionKeys.VIEW);
+		objectEntryService.checkPermission(
+			ActionKeys.VIEW, objectEntry, getUserId());
 
 		return objectEntry;
 	}
@@ -328,20 +408,6 @@ public class ObjectEntryServiceImpl extends ObjectEntryServiceBaseImpl {
 
 		modelResourcePermission.check(
 			getPermissionChecker(), objectEntryId, actionId);
-	}
-
-	private void _checkPermissions(ObjectEntry objectEntry, String actionId)
-		throws PortalException {
-
-		_checkModelResourcePermission(
-			objectEntry.getObjectDefinitionId(), objectEntry.getObjectEntryId(),
-			actionId);
-
-		ObjectEntryPermissionUtil.checkAccountEntryPermission(
-			_accountEntryLocalService, actionId,
-			_objectDefinitionPersistence.findByPrimaryKey(
-				objectEntry.getObjectDefinitionId()),
-			objectEntry, _objectFieldLocalService, getUserId());
 	}
 
 	private void _checkPortletResourcePermission(
